Time: 6 ms (92.58%) | Memory: 71.6 MB (92.06%) - LeetSync

+++ 878-shifting-letters/shifting-letters.java
@@ -1,6 +1,6 @@
 class Solution {
     public String shiftingLetters(String s, int[] shifts) {
-        long suffix=0;
+        int suffix=0;
         char ans[] = s.toCharArray();
 
         for(int i=shifts.length-1;i>=0;i--) {
